+++ src/Network/Network/ChangeLog.md
@@ -19,6 +19,7 @@
 --->
 
 ## Upcoming Release
+* Added private link support on Web/sites, Automation/automationAccounts, Relay/namespaces, Search/searchServices
 
 ## Version 3.2.0
 * Fixed parameters swap in VWan HubVnet connection